refactor: extract foresight document composable
